critical migration fixes

+++ database/migrations/2016_01_18_143320_drop_blogId_from_locations.php
@@ -12,10 +12,13 @@ class DropBlogIdFromLocations extends Migration {
 	 */
 	public function up()
 	{
-		Schema::table('lv_locations', function(Blueprint $table)
+		if(Schema::hasColumn('lv_locations', 'program_id'))
 		{
-			$table->dropColumn('program_id');
-		});
+			Schema::table('lv_locations', function(Blueprint $table)
+			{
+				$table->dropColumn('program_id');
+			});
+		}
 	}
 
 	/**
